+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+# 0.1.11
+    added count*ObjectByClass functions in class iLGE_2D_Engine
+    remove useless "if (object.findElementByType())"
+    remove useless #smartClean in class iLGE_2D_Object
+    improve #smartClean in class iLGE_2D_Engine
+
 # 0.1.10
     added toggle in mouse and keyboard events
-    
+
 # 0.1.9
     implemented z_order in class iLGE_2D_Object
     changed mouse events and pointerLock behavior